removed tests and include in calcVar
