+++ Server/myapi/views.py
@@ -3,8 +3,6 @@ from django.shortcuts import render
 
 from .stockData import get_stock_data
 from . import calcVar
-from . import tests
-
 def getRoutes(request):
 
     US_STOCK_LIST = ["TSM","GOOGL","TSLA","MSFT","AAPL"]
@@ -16,10 +14,10 @@ def getRoutes(request):
 
     print(f'\n\nnumber of share: {number_of_share}')
     print(f'closing price: {closing_price}')
-    weight = tests.get_weight(number_of_share, closing_price)
+    weight = calcVar.get_weight(number_of_share, closing_price)
     print(f'weight: {weight}\n')
 
     calcVar.portfolio(US_STOCK_LIST, weight, 501)
-    
+
     return JsonResponse("123",safe=False)
